Author: Ryan Date: 4/28/2023
Addition of PostgrsqlDBManager and Docker Compsoe for App-DB Deployment

Created minimum database functionality for storing generated forecasts in DB for save and analysis.
Add Dockerfile and compose.yaml for containerization of app and db for easy-deployment to cloud.
Re-built python package distribution and updated requirements.txt and environment.yaml to match.